gpac-agenda/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
pyjwt==2.1.0
bcrypt==3.2.0
python-dateutil==2.8.2
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
